address review: gate lenient constructor behind build tag; fix stale doc comments
seidroid[bot] review on #3691:
- [suggestion] NewTxConfigWithoutBodyBloatRejection was compiled into untagged
  builds, weakening the 'lenient decoder reachable only from historical_replay
  artifacts' invariant. Move it into a //go:build historical_replay file
  (config_historical_replay.go) so an untagged/production binary never compiles
  it — a future caller cannot accidentally wire it into a live path. Default and
  tagged builds both verified.
- [nit] Replace the stale copy-pasted 'amino based test configuration' doc
  comments on MakeEncodingConfig / MakeLegacyEncodingConfig with the lenient/replay purpose.
