+++ sei-cosmos/x/auth/tx/config.go
@@ -41,21 +41,6 @@ func NewTxConfigWithHandler(protoCodec codec.ProtoCodecMarshaler, handler signin
 	}
 }
 
-// NewTxConfigWithoutBodyBloatRejection returns a TxConfig whose decoder does not
-// reject non-canonical (body-bloat) protobuf tx bodies, so historical blocks whose
-// tx bodies predate that check can be decoded and executed. This is consensus-unsafe
-// for live paths and exists only for historical replay tooling gated behind a build tag.
-func NewTxConfigWithoutBodyBloatRejection(protoCodec codec.ProtoCodecMarshaler, enabledSignModes []signingtypes.SignMode) client.TxConfig {
-	return &config{
-		handler:     makeSignModeHandler(enabledSignModes),
-		decoder:     DefaultTxDecoderWithoutBodyBloatRejection(protoCodec),
-		encoder:     DefaultTxEncoder(),
-		jsonDecoder: DefaultJSONTxDecoder(protoCodec),
-		jsonEncoder: DefaultJSONTxEncoder(protoCodec),
-		protoCodec:  protoCodec,
-	}
-}
-
 func (g config) NewTxBuilder() client.TxBuilder {
 	return newBuilder()
 }
